requirements.txt
Django==2.2.6
djangorestframework==3.10.3
drf-writable-nested==0.5.1
object-detection==0.1
pytz==2019.3
sqlparse==0.3.0
